pyproject.toml
[tool.poetry]
name = "playwrighttemplate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.0"
playwright = "^1.29.1"
allure-pytest = "^2.12.0"
pytest-playwright = "^0.3.0"
pytest-xdist = "^3.1.0"
logging = "^0.4.9.6"
pydantic = "^1.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
